Refactor Parquet reader helpers
- Removed `chunked_reader_helpers` and replaced its functionality with `parquet_reader_helpers`.
- Updated includes in various files to reflect the new structure.
- Added new implementations for column path normalization and comparison in `parquet_reader_helpers`.
- Adjusted tests to ensure compatibility with the new helper functions.
