+++ cpp/CMakeLists.txt
@@ -696,8 +696,6 @@ add_library(
   src/io/orc/writer_impl.cu
   src/io/parquet/arrow_schema_writer.cpp
   src/io/parquet/bloom_filter_reader.cu
-  src/io/parquet/chunked_reader_helpers.cpp
-  src/io/parquet/column_path_helpers.cpp
   src/io/parquet/compact_protocol_reader.cpp
   src/io/parquet/compact_protocol_writer.cpp
   src/io/parquet/decode_preprocess.cu
@@ -724,6 +722,7 @@ add_library(
   src/io/parquet/page_string_decode.cu
   src/io/parquet/predicate_pushdown.cpp
   src/io/parquet/reader.cpp
+  src/io/parquet/parquet_reader_helpers.cpp
   src/io/parquet/reader_impl.cpp
   src/io/parquet/reader_impl_chunking.cu
   src/io/parquet/reader_impl_chunking_utils.cu
