+++ cpp/src/io/parquet/experimental/hybrid_scan_helpers.cpp
@@ -5,9 +5,9 @@
 
 #include "hybrid_scan_helpers.hpp"
 
-#include "io/parquet/column_path_helpers.hpp"
 #include "io/parquet/compact_protocol_reader.hpp"
 #include "io/parquet/expression_transform_helpers.hpp"
+#include "io/parquet/parquet_reader_helpers.hpp"
 #include "io/parquet/reader_impl_helpers.hpp"
 
 #include <cudf/detail/nvtx/ranges.hpp>
