+++ cpp/src/io/parquet/expression_transform_helpers.cpp
@@ -5,7 +5,7 @@
 
 #include "expression_transform_helpers.hpp"
 
-#include "column_path_helpers.hpp"
+#include "parquet_reader_helpers.hpp"
 
 #include <cudf/ast/detail/expression_transformer.hpp>
 #include <cudf/ast/detail/operators.hpp>
