+++ cpp/src/io/parquet/parquet_reader_helpers.cpp
@@ -3,11 +3,16 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "column_path_helpers.hpp"
+#include "parquet_reader_helpers.hpp"
+
+#include <cudf/logger.hpp>
+
+#include <cuda/numeric>
 
 #include <algorithm>
 #include <cctype>
 #include <cstddef>
+#include <format>
 #include <functional>
 #include <string>
 #include <string_view>
@@ -49,6 +54,25 @@ column_path_set make_column_path_set(bool case_sensitive_names, std::size_t buck
 {
   return column_path_set(
     bucket_hint, column_path_hash{case_sensitive_names}, column_path_equal{case_sensitive_names});
+}
+
+std::size_t derive_pass_read_limit(std::size_t chunk_read_limit)
+{
+  if (chunk_read_limit == 0) { return 0; }
+
+  // Derive a heuristic pass limit (1.5x the chunk_read_limit) to reduce surprising OOMs
+  auto const sum             = cuda::add_overflow(chunk_read_limit, chunk_read_limit / 2);
+  auto const pass_read_limit = sum.overflow ? 0 : sum.value;
+
+  CUDF_LOG_WARN(std::format(
+    "Chunked Parquet reader: a chunk_read_limit ({} bytes) was provided without a "
+    "pass_read_limit; defaulting pass_read_limit to {} bytes to bound input and decompression "
+    "memory and reduce the risk of out-of-memory errors on large files. Use a constructor overload "
+    "that accepts pass_read_limit to control this explicitly.",
+    chunk_read_limit,
+    pass_read_limit));
+
+  return pass_read_limit;
 }
 
 }  // namespace cudf::io::parquet::detail
