+++ cpp/src/io/parquet/parquet_reader_helpers.hpp
@@ -100,4 +100,17 @@ template <typename Value>
     bucket_hint, column_path_hash{case_sensitive_names}, column_path_equal{case_sensitive_names});
 }
 
+/**
+ * @brief Derives a bounded input `pass_read_limit` from a `chunk_read_limit`.
+ *
+ * Used by the `chunked_parquet_reader` constructor overloads that do not accept a `pass_read_limit`
+ * to bound input and decompression memory and reduce the risk of out-of-memory errors on large
+ * files. A `chunk_read_limit` of 0 means unlimited input memory and is returned unchanged;
+ * otherwise a heuristic limit is derived and a warning is emitted.
+ *
+ * @param chunk_read_limit The output chunk byte limit
+ * @return The derived input pass byte limit
+ */
+[[nodiscard]] std::size_t derive_pass_read_limit(std::size_t chunk_read_limit);
+
 }  // namespace cudf::io::parquet::detail
